FT: Implementado realizar faxina e repor produto

+++ lib/pages/cleaning/models/cleaning_repository.dart
@@ -3,6 +3,9 @@ import 'package:faxinapp/pages/cleaning/models/cleaning.dart';
 import 'package:faxinapp/pages/products/models/product.dart';
 import 'package:faxinapp/pages/tasks/models/task.dart';
 
+import 'cleaning_product.dart';
+import 'cleaning_task.dart';
+
 class CleaningRepository {
   static final CleaningRepository _cleaningRepository =
       CleaningRepository._internal(AppDatabase.get());
@@ -17,16 +20,17 @@ class CleaningRepository {
 
   Future<bool> save(Cleaning cleaning) async {
     var db = await _appDatabase.getDb();
-    
+
     if (cleaning.id == null || cleaning.id == 0) {
       cleaning.id = await db.insert(CleaningTable.table, {
         CleaningTable.NAME: cleaning.name,
         CleaningTable.GUIDELINES: cleaning.guidelines,
         CleaningTable.FREQUENCY: cleaning.frequency.index,
         CleaningTable.NEXT_DATE: cleaning.nextDate.toIso8601String(),
-        CleaningTable.START_DATE: cleaning.startDate != null ? cleaning.startDate.toIso8601String() : null,
-        CleaningTable.END_DATE: cleaning.endDate != null ? cleaning.endDate.toIso8601String() : null,
-        CleaningTable.ESTIMATED_TIME : cleaning.estimatedTime.toString()
+        CleaningTable.DUE_DATE: cleaning.dueDate != null
+            ? cleaning.dueDate.toIso8601String()
+            : null,
+        CleaningTable.ESTIMATED_TIME: cleaning.estimatedTime.toString()
       });
     } else {
       await db.update(
@@ -36,15 +40,16 @@ class CleaningRepository {
             CleaningTable.GUIDELINES: cleaning.guidelines,
             CleaningTable.FREQUENCY: cleaning.frequency.index,
             CleaningTable.NEXT_DATE: cleaning.nextDate.toIso8601String(),
-            CleaningTable.START_DATE: cleaning.startDate.toIso8601String(),
-            CleaningTable.END_DATE: cleaning.endDate.toIso8601String(),
-            CleaningTable.ESTIMATED_TIME :  cleaning.estimatedTime.toString()
+            CleaningTable.DUE_DATE: cleaning.dueDate != null
+                ? cleaning.dueDate.toIso8601String()
+                : null,
+            CleaningTable.ESTIMATED_TIME: cleaning.estimatedTime.toString()
           },
           where: '${CleaningTable.ID} = ? ',
           whereArgs: [cleaning.id]);
     }
 
-    if (cleaning.products.isNotEmpty) {
+    if (cleaning.products != null && cleaning.products.isNotEmpty) {
       await db.delete(CleaningProductTable.table,
           where: '${CleaningProductTable.REF_CLEANING} = ?',
           whereArgs: [cleaning.id]);
@@ -87,11 +92,49 @@ class CleaningRepository {
         where: '${CleaningProductTable.REF_CLEANING} = ?', whereArgs: [c.id]);
   }
 
-  Future<List<Cleaning>> findAll() async {
+  Future<Cleaning> find(int id) async {
     var db = await _appDatabase.getDb();
 
-    var result =
-        await db.query(CleaningTable.table, orderBy: CleaningTable.NAME);
+    var result = await db.query(CleaningTable.table,
+        where: '${CleaningTable.ID} = ?',
+        whereArgs: [id],
+        orderBy: CleaningTable.NAME);
+
+    if (result.isNotEmpty) {
+      Cleaning c = Cleaning.fromMap(result.first);
+
+      var fetchProducts = await db.query(ProductTable.table,
+          where: "${ProductTable.ID} in ( " +
+              "SELECT ${CleaningProductTable.REF_PRODUCT} FROM ${CleaningProductTable.table} WHERE  ${CleaningProductTable.REF_CLEANING} = ? )",
+          whereArgs: [c.id]);
+
+      c.products = [];
+      for (Map<String, dynamic> p in fetchProducts) {
+        c.products.add(Product.fromMap(p));
+      }
+
+      var fetchTasks = await db.query(TaskTable.table,
+          where:
+              "${TaskTable.ID} in ( SELECT ${CleaningTaskTable.REF_TASK} FROM ${CleaningTaskTable.table} WHERE  ${CleaningTaskTable.REF_CLEANING} = ? )",
+          whereArgs: [c.id]);
+
+      c.tasks = [];
+      for (Map<String, dynamic> t in fetchTasks) {
+        c.tasks.add(Task.fromMap(t));
+      }
+
+      return c;
+    }
+
+    return null;
+  }
+
+  Future<List<Cleaning>> findPendents() async {
+    var db = await _appDatabase.getDb();
+
+    var result = await db.query(CleaningTable.table,
+        where: '${CleaningTable.DUE_DATE} is null',
+        orderBy: CleaningTable.NEXT_DATE);
 
     List<Cleaning> cleanings = List();
 
@@ -122,5 +165,95 @@ class CleaningRepository {
     }
 
     return cleanings;
+  }
+
+  Future<List<Cleaning>> findAll() async {
+    var db = await _appDatabase.getDb();
+
+    var result =
+        await db.query(CleaningTable.table, orderBy: CleaningTable.NEXT_DATE);
+
+    List<Cleaning> cleanings = List();
+
+    for (Map<String, dynamic> item in result) {
+      Cleaning c = Cleaning.fromMap(item);
+
+      var fetchProducts = await db.query(ProductTable.table,
+          where: "${ProductTable.ID} in ( " +
+              "SELECT ${CleaningProductTable.REF_PRODUCT} FROM ${CleaningProductTable.table} WHERE  ${CleaningProductTable.REF_CLEANING} = ? )",
+          whereArgs: [c.id]);
+
+      c.products = [];
+      for (Map<String, dynamic> p in fetchProducts) {
+        c.products.add(Product.fromMap(p));
+      }
+
+      var fetchTasks = await db.query(TaskTable.table,
+          where:
+              "${TaskTable.ID} in ( SELECT ${CleaningTaskTable.REF_TASK} FROM ${CleaningTaskTable.table} WHERE  ${CleaningTaskTable.REF_CLEANING} = ? )",
+          whereArgs: [c.id]);
+
+      c.tasks = [];
+      for (Map<String, dynamic> t in fetchTasks) {
+        c.tasks.add(Task.fromMap(t));
+      }
+
+      cleanings.add(c);
+    }
+
+    return cleanings;
+  }
+
+  Future finish(Cleaning cleaning, List<CleaningTask> tasks,
+      List<CleaningProduct> products) async {
+    var db = await _appDatabase.getDb();
+
+    //finaliza a limpeza
+    await db.update(CleaningTable.table,
+        {CleaningTable.DUE_DATE: DateTime.now().toIso8601String()},
+        where: '${CleaningTable.ID} = ? ', whereArgs: [cleaning.id]);
+
+    tasks.forEach((task) async => await db.update(
+        CleaningTaskTable.table, {CleaningTaskTable.REALIZED: task.realized},
+        where:
+            "${CleaningTaskTable.REF_CLEANING} = ? and ${CleaningTaskTable.REF_TASK} = ?",
+        whereArgs: [cleaning.id, task.task.id]));
+
+    products.forEach((product) async {
+      await db.update(
+          CleaningProductTable.table,
+          {
+            CleaningProductTable.REALIZED: product.realized,
+            CleaningProductTable.AMOUNT: product.amount
+          },
+          where:
+              "${CleaningProductTable.REF_CLEANING} = ? and ${CleaningProductTable.REF_PRODUCT} = ?",
+          whereArgs: [cleaning.id, product.product.id]);
+
+      await db.update(
+          ProductTable.table,
+          {
+            ProductTable.CURRENT_CAPACITY:
+                product.product.currentCapacity - product.amount,
+          },
+          where: "${ProductTable.ID} = ?",
+          whereArgs: [product.product.id]);
+    });
+
+    products.forEach((product) => {});
+
+    //cria nova a limpeza
+    Cleaning newCleaning = Cleaning();
+    newCleaning.estimatedTime = cleaning.estimatedTime;
+    newCleaning.frequency = cleaning.frequency;
+    newCleaning.guidelines = cleaning.guidelines;
+    newCleaning.name = cleaning.name;
+    newCleaning.tasks = cleaning.tasks;
+    newCleaning.products = cleaning.products;
+    newCleaning.nextDate = cleaning.futureDate();
+
+    await this.save(newCleaning);
+
+    return newCleaning;
   }
 }
